+++ src/components/LinkItems/Section.vue
@@ -17,9 +17,9 @@
       :style="gridStyle"
     >
       <Item
-        v-for="(item, index) in items"
-        :id="`${index}_${makeId(item.title)}`"
-        :key="`${index}_${makeId(item.title)}`"
+        v-for="(item) in sortedItems"
+        :id="makeId(title, item.title)"
+        :key="makeId(title, item.title)"
         :url="item.url"
         :title="item.title"
         :description="item.description"
@@ -31,6 +31,7 @@
         :statusCheckHeaders="item.statusCheckHeaders"
         :itemSize="newItemSize"
         :hotkey="item.hotkey"
+        :provider="item.provider"
         :enableStatusCheck="shouldEnableStatusCheck(item.statusCheck)"
         :statusCheckInterval="getStatusCheckInterval()"
         :statusCheckAllowInsecure="item.statusCheckAllowInsecure"
@@ -49,6 +50,8 @@
 </template>
 
 <script>
+import { sortOrder as defaultSortOrder, localStorageKeys } from '@/utils/defaults';
+import ErrorHandler from '@/utils/ErrorHandler';
 import Item from '@/components/LinkItems/Item.vue';
 import Collapsable from '@/components/LinkItems/Collapsable.vue';
 import IframeModal from '@/components/LinkItems/IframeModal.vue';
@@ -71,6 +74,28 @@ export default {
     IframeModal,
   },
   computed: {
+    sortOrder() {
+      return this.displayData.sortBy || defaultSortOrder;
+    },
+    /* If the sortBy attribute is specified, then return sorted data */
+    sortedItems() {
+      let { items } = this;
+      if (this.config.appConfig.disableSmartSort) return items;
+      if (this.sortOrder === 'alphabetical') {
+        this.sortAlphabetically(items);
+      } else if (this.sortOrder === 'reverse-alphabetical') {
+        this.sortAlphabetically(items).reverse();
+      } else if (this.sortOrder === 'most-used') {
+        items = this.sortByMostUsed(items);
+      } else if (this.sortOrder === 'last-used') {
+        items = this.sortBLastUsed(items);
+      } else if (this.sortOrder === 'random') {
+        items = this.sortRandomly(items);
+      } else if (this.sortOrder && this.sortOrder !== 'default') {
+        ErrorHandler(`Unknown Sort order '${this.sortOrder}' under '${this.title}'`);
+      }
+      return items;
+    },
     newItemSize() {
       return this.displayData.itemSize || this.itemSize;
     },
@@ -89,8 +114,9 @@ export default {
   },
   methods: {
     /* Returns a unique lowercase string, based on name, for section ID */
-    makeId(str) {
-      return str.replace(/\s+/g, '-').replace(/[^a-zA-Z ]/g, '').toLowerCase();
+    makeId(sectionStr, itemStr) {
+      const charSum = sectionStr.split('').map((a) => a.charCodeAt(0)).reduce((x, y) => x + y);
+      return `${charSum}_${itemStr.replace(/\s+/g, '-').replace(/[^a-zA-Z ]/g, '').toLowerCase()}`;
     },
     /* Opens the iframe modal */
     triggerModal(url) {
@@ -105,12 +131,38 @@ export default {
       const globalPreference = this.config.appConfig.statusCheck || false;
       return itemPreference !== undefined ? itemPreference : globalPreference;
     },
+    /* Determine how often to re-fire status checks */
     getStatusCheckInterval() {
       let interval = this.config.appConfig.statusCheckInterval;
       if (!interval) return 0;
       if (interval > 60) interval = 60;
       if (interval < 1) interval = 0;
       return interval;
+    },
+    /* Sorts items alphabetically using the title attribute */
+    sortAlphabetically(items) {
+      return items.sort((a, b) => (a.title > b.title ? 1 : -1));
+    },
+    /* Sorts items by most used to least used, based on click-count */
+    sortByMostUsed(items) {
+      const usageCount = JSON.parse(localStorage.getItem(localStorageKeys.MOST_USED) || '{}');
+      const gmu = (item) => usageCount[this.makeId(this.title, item.title)] || 0;
+      items.reverse().sort((a, b) => (gmu(a) < gmu(b) ? 1 : -1));
+      return items;
+    },
+    /* Sorts items by most recently used */
+    sortBLastUsed(items) {
+      const usageCount = JSON.parse(localStorage.getItem(localStorageKeys.LAST_USED) || '{}');
+      const glu = (item) => usageCount[this.makeId(this.title, item.title)] || 0;
+      items.reverse().sort((a, b) => (glu(a) < glu(b) ? 1 : -1));
+      return items;
+    },
+    /* Sorts items randomly */
+    sortRandomly(items) {
+      return items
+        .map((value) => ({ value, sort: Math.random() }))
+        .sort((a, b) => a.sort - b.sort)
+        .map(({ value }) => value);
     },
   },
 };